Remove temporary patch for helm
We were not able to delete multiple markers due to restriction of helm.
This was temporary fixed by patching with `flet`, but finally helm fixed
this issue.

For more details, see
https://github.com/emacs-helm/helm/issues/706#issuecomment-64076668

Note that helm uses `equal` for equality checking of REAL of
(DISPLAY . REAL) pairs.  Jumar uses `jumar:node` for REAL.  This will
cause a bug if we implement reference count of `jumar:jumarker`
because `*jumar:jm-tree*` and `*jumar:jm-list*` can have similar
structure under the use of `'both`.  Currently, this is not a real problem
because they have different Emacs marker objects, even they are almost
the same.  See `'both` case of `jumar-dwin-add-marker`.